fix(rpc): wrong import path for moved files
